requirements.txt
discord.py==1.5.0
loguru==0.5.3
psycopg2==2.8.6